[push] removes protocol buffers
